# PERF: GroupBy.count