```
from asv_bench.benchmarks.groupby import *
self = GroupByMethods()
self.setup("datetime", "count", "direct", 2)

%timeit self.time_dtype_as_field('datetime', 'count', 'direct', 2)
73.7 µs ± 982 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- master
48.7 µs ± 473 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR
```